restructured templates standalone
