+++ apps/Standalone/src/templates/app/TemplatesStandaloneDesigner.tsx
@@ -1,40 +1,32 @@
+import type { ReactNode } from 'react';
+import { getReactQueryClient } from '@microsoft/logic-apps-designer';
+import { QueryClientProvider, useQuery } from 'react-query';
+import { loadToken } from '../../environments/environment';
 import { DevToolbox } from '../components/DevToolbox';
 import type { RootState } from '../state/Store';
-import { AzureThemeDark } from '@fluentui/azure-themes/lib/azure/AzureThemeDark';
-import { AzureThemeLight } from '@fluentui/azure-themes/lib/azure/AzureThemeLight';
-import { ThemeProvider } from '@fluentui/react';
-import { FluentProvider, webDarkTheme, webLightTheme } from '@fluentui/react-components';
-import { PortalCompatProvider } from '@fluentui/react-portal-compat';
 import { TemplatesDataProvider, TemplatesDesigner, TemplatesDesignerProvider } from '@microsoft/logic-apps-templates';
-import { Theme as ThemeType } from '@microsoft/logic-apps-shared';
 import { useSelector } from 'react-redux';
 
+const standaloneQueryClient = getReactQueryClient();
+
+const LoadWhenArmTokenIsLoaded = ({ children }: { children: ReactNode }) => {
+  const { isLoading } = useQuery('armToken', loadToken);
+  return isLoading ? null : <>{children}</>;
+};
 export const TemplatesStandaloneDesigner = () => {
   const theme = useSelector((state: RootState) => state.templateDataLoader.theme);
   const currentTemplate = useSelector((state: RootState) => state.templateDataLoader.currentTemplate);
-  const isLightMode = theme === ThemeType.Light;
 
   return (
-    <div style={{ flex: '1 1 1px', display: 'flex', flexDirection: 'column' }}>
-      <div style={{ flex: '0 1 1px' }}>
-        <ThemeProvider theme={isLightMode ? AzureThemeLight : AzureThemeDark}>
-          <FluentProvider theme={isLightMode ? webLightTheme : webDarkTheme}>
-            {/* eslint-disable-next-line @typescript-eslint/ban-ts-comment */}
-            {/* @ts-ignore */}
-            <PortalCompatProvider>
-              <DevToolbox />
-            </PortalCompatProvider>
-          </FluentProvider>
-        </ThemeProvider>
-      </div>
-
-      <div style={{ flex: '1 1 1px', display: 'flex', flexDirection: 'column' }}>
+    <QueryClientProvider client={standaloneQueryClient}>
+      <LoadWhenArmTokenIsLoaded>
+        <DevToolbox />
         <TemplatesDesignerProvider locale="en-US" theme={theme} options={{}}>
           <TemplatesDataProvider currentTemplate={currentTemplate} theme={theme}>
             <TemplatesDesigner />
           </TemplatesDataProvider>
         </TemplatesDesignerProvider>
-      </div>
-    </div>
+      </LoadWhenArmTokenIsLoaded>
+    </QueryClientProvider>
   );
 };
